In commit fd09a48563a36752cdf9f0685e586e7869683d51 I added a line to the grid constructor which sorted the velocity field in increasing order, since the algorithm of Mihalas (1980) requires monotonic fields.
But then I realized that the explosion model data itself should already be
monotonic, and if it's not, we need the program to exit because that's a
physical inconsistency that must be resolved "upstream" of the radiative
transfer calculation.

So in 958fffcbcf2a4fc9716f5071aaa790f0b9928248 I removed that line. However, the
example hydro model is one from Peter Hoeflich, and while his data are indeed
monotonic, they're also reversed. I forgot to reverse them back, so the
interpolator started crashing because it requires vectors on which to
interpolate to the monotonically *increasing*, not *decreasing.* So here I've
added a "reverse" line which flips the velocity field over.