Migrating build system to Autotools.
CMake is certainly easier to use for simple projects - the "time to
science" is lower than that of Autotools. But eventually I realized that
CMake actually a huge pain to use because a.) the documentation is
*abysmal*, and b.) it uses an enormous number of "local" environment
variables under the covers and then cobbles them all together, which
makes managing, e.g., the include flags or compiler optimization
settings for certain files, nearly impossible. Autotools, on the other
hand, is challenging to learn, but immensely more powerful once you've
figured out how to use it.